refactoring app logging
