+++ test/unit/app/logging/backend-logging-transport.spec.js
@@ -18,8 +18,7 @@
 // @flow
 
 import { describe, it, expect, beforeEach } from '../../../helpers/dependencies'
-import { BackendLogCachingTransport }
-  from '../../../../src/app/logging/backend-logging-transports'
+import BackendLogCachingTransport from '../../../../src/app/logging/backend-log-transport'
 import LogCache from '../../../../src/app/logging/log-cache'
 
 describe('BackendLogCachingTransport', () => {
